+++ CHANGES.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Add 83 new playable maps
 * Add debug setting to allow loading already completed/won/lost islands
 * Add cheat on F9 which sets half the hexagons to the current team
   * Used to test real AI surrendering